fix(result): move all styles and scripts to result head (only keep markup and editor script)
this prevents setting of document.body.innerHTML from preventing further updates to the page
